fix https://github.com/huggingface/datasets/issues/2032

benchmarking is somewhat inconclusive, currently running on `book_corpus` to get a better sense of the improvement

```diff
3,11c3,11
<   "map identity": 16.943350610003108,
<   "map identity batched": 0.9310493359953398,
<   "map no-op batched": 0.632951049003168,
<   "map no-op batched numpy": 0.6800740210019285,
<   "map no-op batched pandas": 0.4896433609974338,
<   "map no-op batched pytorch": 0.5659142479998991,
<   "map no-op batched tensorflow": 1.3155841589905322,
<   "map fast-tokenizer batched": 14.161769978993107,
<   "filter": 9.384816927995416
---
>   "map identity": 16.720537469998817,
>   "map identity batched": 0.909121311007766,
>   "map no-op batched": 0.5962835579994135,
>   "map no-op batched numpy": 0.6866216960042948,
>   "map no-op batched pandas": 0.4974203809979372,
>   "map no-op batched pytorch": 0.5545022570004221,
>   "map no-op batched tensorflow": 1.3693314480042318,
>   "map fast-tokenizer batched": 13.003258584998548,
>   "filter": 9.322257651001564
```